// src/app/api/models/social-media-generation-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SocialMediaGenerationResponse {
  aiDetectionScore?: ({
'aiGenerated': number;
'aiParaphrased': number;
'humanWritten': number;
'showsAiSigns': number;
}) | null;
  analysisReportKey?: string | null;
  analysisReportUrl?: string | null;
  createdAt: string;
  generatedPosts: {
[key: string]: {
'body': string;
'hashtags': Array<string>;
'emojis'?: string;
'hook'?: string;
'image'?: {
'url'?: string;
'r2Key'?: string;
'mimeType'?: string;
};
};
};
  id: string;
  language: string | null;
  networks: {
[key: string]: boolean;
};
  niche: string;
  r2Key?: string | null;
  roiScore?: number | null;
  topicDescription: string | null;
  topicTitle: string;
  userId: string;
  viralityScore?: number | null;
}